Add a "hiddenTags" Params field
All tags listed in hiddenTags aren't displayed on the posts.
Tags can be used to categorize posts, and categories don't need to be
displayed.

An example is if the pages are used for about-us and contact, and
posts as blog articles and projects presentation.

Example of hugo.toml using this new field:

[params]
hiddenTags = ["project", "blog"]
[menu]
[[menu.main]]
  identifier = "projects"
  url = "/tags/project/"
  name = "My Projects"
  weight = 1
[[menu.main]]
  identifier = "blog"
  url = "/tags/blog/"
  name = "Blog"
  weight = 2